% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.Orientation.Army.R
\name{Convert}
\alias{Convert}
\title{Convert}
\usage{
Convert(jumping = 0, ball_toss = 0, equilibrium = 0, planking = 0, running = 0)
}
\arguments{
\item{jumping}{correspond to the homonymous exercise score. The default value is 0. The measurement system is in meters}

\item{ball_toss}{correspond to the homonymous exercise score. The default value is 0.The measurement system is in meters}

\item{equilibrium}{correspond to the homonymous exercise score. The default value is 0. The measurement system is in seconds}

\item{planking}{correspond to the homonymous exercise score. The default value is 0. The measurement system is in seconds}

\item{running}{correspond to the homonymous exercise score. The default value is 0. The measurement system is in seconds}
}
\value{
A \code{list} of points:
\describe{
    \item{list}{The order of the points correspond to the following exercises result: "jumping" , "ball toss" , "equilibrium" , "planking" , "running" }
}
}
\description{
This function "Convert" will first load the table of points, then will use the score of the army exercises to convert them into a \code{list} of point. Also, the function will \code{plot} a radar-chart showing the correlated points.
}
\examples{
Convert(jumping = 2, ball_toss = 5, equilibrium = 30, planking = 50, running = 300)
}
\author{
Group-E
}
