% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.Orientation.Army.R
\name{militaryapp}
\alias{militaryapp}
\title{Military App}
\usage{
militaryapp()
}
\value{
The shiny app of Group-E
}
\description{
This will recall the Shiny-App created by Group-E regarding "Performance Scoreboard for Recruitment"
}
\author{
Group-E
}
