% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.Orientation.Army.R
\name{Orientation}
\alias{Orientation}
\title{Orientation}
\usage{
Orientation(results)
}
\arguments{
\item{results}{it has to be a \code{list} of numerical elements}
}
\value{
A \code{df} of the possible weapons class based on the sum of the points used from the input list.
\describe{ The First column represent the name of the army class and the Second column the minimum point to apply for the class.
}
}
\description{
This function will use the result generated by the "Convert" function to create a data-frame listing weapons class based on the sum of point from the list created before. The point used as an input will be all between 0 and 25 thanks to the output of the "Convert" function.
}
\examples{
Orientation(list(5, 10, 20, 19, 25))
}
\author{
Group-E
}
