% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.Orientation.Army.R
\docType{data}
\name{Performances}
\alias{Performances}
\title{Sport Performances}
\format{
## Performances
A data frame with 19 rows and 2 columns:
\describe{
  \item{Fonction}{The Class Army}
  \item{min nb points}{Minimum points to reach the class}
}
}
\source{
load("data/Performance_sportive_et_nb_points.rda")
}
\usage{
Performances
}
\description{
Sport Performances
}
\keyword{datasets}
