% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.Orientation.Army.R
\docType{data}
\name{points_army}
\alias{points_army}
\title{Points Score Data}
\format{
## points_army
A data frame with 26 rows and 6 columns:
\describe{
  \item{points}{Points referring to the score}
  \item{jumping}{Levels of jump scores}
  \item{ball_toss}{Levels of ball tosss}
  \item{equilibrium}{Levels of equilibrium}
  \item{planking}{Levels of planking}
  \item{running}{Levels of running}
}
}
\source{
load("data/points_army.rda")
}
\usage{
points_army
}
\description{
Points Score Data
}
\keyword{datasets}
